Getting rid of the excessive `Token_Endpoints` class.
